adding comments , placing useful functions in a separate file , adding high price less clients feature
